requirements.txt
aardwolf==0.2.2
adblockparser==0.7
aesedb==0.1.3
aiocmd==0.1.2
aioconsole==0.6.2
aiodns==3.0.0
aiofiles==23.2.1
aiohttp==3.8.6
aiomultiprocess==0.9.0
aioquic==0.9.21
aiosignal==1.3.1
aiosmb==0.4.4
aiosqlite==0.17.0
aiowinreg==0.0.7
ajpy==0.0.5
altgraph==0.17.4
aniso8601==9.0.1
anyio==3.7.0
appdirs==1.4.4
arc4==0.3.0
argcomplete==3.1.4
asciitree==0.3.3
asgiref==3.7.2
asn1crypto==1.5.1
asn1tools==0.164.0
asttokens==2.4.1
asyauth==0.0.9
async-timeout==4.0.3
asysocks==0.2.2
attrs==23.1.0
autocommand==2.2.2
Automat==22.10.0
awscli==2.12.0
awscrt==1.0.0.dev0
Babel==2.10.3
backcall==0.2.0
backoff==2.2.1
bcrypt==3.2.2
beautifulsoup4==4.12.2
beniget==0.4.1
bidict==0.22.1
binwalk==2.3.3
bitstruct==8.15.1
blinker==1.7.0
bluepy==1.3.0
Brlapi==0.8.5
Brotli==1.1.0
catfish==4.16.4
censys==2.2.9
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
cheroot==10.0.0+ds1
CherryPy==18.8.0
cherrypy-cors==1.6
cliapp==1.20180812.1
click==8.1.6
click-plugins==1.1.1
cmdtest==0.32+git
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.7
crackmapexec==5.4.0
cryptography==38.0.4
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
debtags==2.1
decorator==5.1.1
Deprecated==1.2.14
dicttoxml==1.7.15
diskcache==5.4.0
distlib==0.3.7
distro==1.8.0
distro-info==1.7
Django==4.2.3
django-cors-headers==4.3.1
django-tailwindcss==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnslib==0.9.23
dnspython==2.4.2
docopt==0.6.2
docutils==0.20.1
donut-shellcode==0.9.3
dropbox==11.36.2
dsinternals==1.2.4
ecdsa==0.18.0
et-xmlfile==1.0.1
executing==1.2.0
ExifRead==3.0.0
fastapi==0.101.0
fierce==1.5.0
filelock==3.12.4
flasgger==0.9.5
Flask==2.2.5
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.2
fonttools==4.46.0
frozenlist==1.4.0
fs==2.4.16
future==0.18.2
gast==0.5.2
GDAL==3.8.1
gpg==1.18.0
greenlet==2.0.2
gyp==0.1
h11==0.14.0
h2==4.1.0
hashID==3.1.4
hiredis==1.0.1
hpack==4.0.0
html5lib==1.1
httpagentparser==1.9.1
httpcore==0.17.3
httplib2==0.20.4
httpx==0.23.3
humanize==4.8.0
hyperframe==6.0.0
hyperlink==21.0.0
idna==3.3
impacket==0.11.0
importlib-metadata==4.12.0
importlib-resources==6.0.1
incremental==21.3.0
inflect==7.0.0
iniconfig==1.1.1
invoke==2.0.0
ipwhois==1.2.0
IPy==1.1
ipython==8.14.0
itsdangerous==2.1.2
jaraco.collections==4.2.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.text==3.11.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jq==1.2.1
jsonpointer==2.3
jsonschema==4.10.3
kaitaistruct==0.10
kali-tweaks==2023.3.2
KismetCaptureBtGeiger==2021.7.1
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2020.10.1
KismetCaptureRtladsb==2020.10.1
KismetCaptureRtlamr==2020.10.1
kiwisolver==0.0.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
ldap3==2.9.1
ldapdomaindump==0.9.4
lightdm-gtk-greeter-settings==1.2.2
limits==3.6.0
louis==3.27.0
lsassy==3.1.6
lxml==4.9.3
lz4==4.0.2+dfsg
macholib==1.16.3
Mako==1.2.4.dev0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
masky==0.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mechanize==0.4.9
minidump==0.0.21
minikerberos==0.4.0
mistune0==0.8.4
mitmproxy==10.1.5
more-itertools==10.1.0
mpmath==0.0.0
msgpack==1.0.3
msldap==0.4.7
multidict==6.0.4
mysqlclient==1.4.6
nassl==5.0.1
neo4j==5.2.dev0
neobolt==1.7.17
neotime==1.7.4
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.8
numpy==1.24.2
oauthlib==3.2.2
olefile==0.46
onboard==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.2
paramiko==2.12.0
parso==0.8.3
passlib==1.7.4
Paste==3.7.1
PasteDeploy==3.0.1
PasteScript==3.2.1
patator==1.0
pcapy==0.11.5.dev0
pexpect==4.8.0
phonenumbers==8.12.57
pickleshare==0.7.5
Pillow==6.2.1
platformdirs==4.1.0
pluggy==1.3.0
pluginbase==1.0.1
ply==3.11
portend==3.1.0
prettytable==3.6.0
prompt-toolkit==3.0.41
protobuf==4.21.12
psycopg2==2.9.9
ptyprocess==0.7.0
publicsuffix2==2.20191221
publicsuffixlist==0.9.3
pure-eval==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.25.1
pycares==4.3.0
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.45.2
pydantic==1.10.4
PyDispatcher==2.0.5
pydot==1.4.2
pyee==9.0.4
pygame==2.5.2
pygexf==0.2.2
Pygments==2.15.1
PyGObject==3.46.0
pygraphviz==1.7
PyHamcrest==2.0.3
pyinotify==0.9.6
PyInstaller==3.5+498e6ee058
PyJWT==2.7.0
pylnk3==0.4.2
pylsqpack==0.3.17
pymssql==2.2.2
PyMySQL==1.0.2
PyNaCl==1.5.0
PyOpenGL==3.1.6
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==2.12.1
pyperclip==1.8.2
pyppeteer==1.0.1
pypsrp==0.8.1
pypykatz==0.6.6
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQt6==6.6.0
PyQt6-sip==13.6.0
pyqtgraph==0.13.3
pyrsistent==0.18.1
PySecretSOCKS==0.9.1
pyserial==3.5
pysmbc==1.0.23
pysmi==0.3.2
pysnmp==4.4.12
PySocks==1.7.1
pyspnego==0.8.0
pytest==7.4.3
python-apt==2.7.0+b1
python-dateutil==2.8.2
python-debian==0.1.49
python-docx==1.1.0
python-dotenv==1.0.0
python-engineio==4.3.4
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.5
python-pam==2.0.2
python-pptx==0.6.18
python-socketio==5.7.2
pythran==0.14.0
pytz==2023.3.post1
pyVNC==0.1
pywerview==0.3.3
pyxdg==0.28
PyYAML==6.0.1
redis==4.3.4
repoze.lru==0.7
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
retrying==1.3.3
rfc3986==2.0.0
rfc3987==1.3.8
rich==13.3.1
roman==3.3
Routes==2.5.1
rq==1.15.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scapy==2.5.0
scipy==1.10.1
sdkmanager==0.6.5
secure==0.3.0
service-identity==23.1.0
shodan==1.30.1
simplejson==3.19.2
six==1.16.0
slowapi==0.1.4
smbmap==1.9.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.47
SQLAlchemy-Utc==0.14.0
sqlparse==0.4.4
sslyze==5.2.0
stack-data==0.6.2
starlette==0.31.1
stone==3.3.1
sympy==1.12
Tempita==0.5.2
tempora==5.5.0
termcolor==1.1.0
terminaltables==3.1.10
theHarvester==4.4.4
tldextract==3.1.2
tls-parser==1.2.2
tomli==2.0.1
tornado==6.3.2
tqdm==4.64.1
traitlets==5.5.0
ttystatus==0.38
Twisted==22.4.0
typing_extensions==4.7.1
ufoLib2==0.16.0
ujson==5.8.0
unicodecsv==0.14.1
unicodedata2==15.1.0
unicrypto==0.0.10
uritemplate==4.1.1
urllib3==1.26.18
urwid==2.1.2
uuid==1.30
uvicorn==0.24.0
uvloop==0.19.0
virtualenv==20.25.0+ds
wadllib==1.3.6
wafw00f==2.2.0
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.2.3
websockets==10.4
websockify==0.10.0
Werkzeug==2.2.2
wfuzz==3.1.0
whois==0.8
wifite==2.7.0
winacl==0.1.7
wrapt==1.15.0
wsproto==1.2.0
xdg==5
xlrd==2.0.1
XlsxWriter==3.0.2
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.13.0
yara-python==4.3.1
yarl==1.8.2
zc.lockfile==2.0
zipp==1.0.0
zlib-wrapper==0.1.3
zope.interface==5.5.2
zstandard==0.22.0
